+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ## [v2.0.17] / 2025-07-xx
 - Bugfix: Use "WindowHandle" as the unique window id for Win32, as that turns out to be determinate. Also fixes the "Windows" page in the GUI.
+- Refactor: Windows - Make foreground window tracking a bunch more efficient.
+- Feature: Windows - "MainWindow": Match on whether a window is a process's "main" window.
 
 ## [v2.0.16] / 2025-07-05
 - Bugfix: Implemented missing parts for "ArgumentList" to work.